requirements.txt
agno
python-dotenv
packaging
uvicorn
groq
google-generativeai
flask
googlesearch-python 
pycountry
fastapi[standard]
pywa[fastapi]
exa-py
SQLAlchemy